# Fix mis-copy of chunk in make_with_timestamps