Signed-off-by: Li Liu <li.liu@zilliz.com>

issue: #19971 

`get_chunk` will return a `const Chunk&`. If we use `auto x = get_chunk()`, `auto` will be deducted to `Chunk` instead of `const Chunk&`. This will cause extra copy of chunks.

To try get the type of a variable, we can do: 
```
#include <boost/type_index.hpp>
using boost::typeindex::type_id_with_cvr;
std::cout << type_id_with_cvr<decltype(x)>().pretty_name() << std::endl;
```

For 6M SIFT data, this copy can add ~1-1.5 ms to the latency for each segment search.
#### Improvement
With this fix, we can gain some benefit on the performance. In my local test environment(3c8g):
**Serial**: 
latency: 11ms->6ms
**Parallel**: 
latency: 49ms->37ms 
qps: 200->260